# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Sistema Bancário API",
        "description": "API completa para sistema bancário com operações de contas, transações, cartões, seguros e relatórios",
        "contact": {
            "name": "Suporte Técnico",
            "email": "admin@banco.ao"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Servidor de Desenvolvimento"
        }
    ],
    "paths": {
...
        "/api/relatorios/transacoes": {
            "get": {
                "tags": [
                    "Relatórios"
                ],
                "summary": "Relatório detalhado de transações",
                "operationId": "0cd9f5bfbae7e29d60f3b1294459ab83",
                "parameters": [
                    {
                        "name": "data_inicio",
                        "in": "query",
                        "description": "Data início",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "data_fim",
                        "in": "query",
                        "description": "Data fim",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "tipo",
                        "in": "query",
                        "description": "Tipo de transação",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "moeda",
                        "in": "query",
                        "description": "Código da moeda",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Itens por página",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Relatório de transações"
                    }
                }
            }
        },
        "/api/relatorios/contas/{id}/extrato": {
            "get": {
                "tags": [
                    "Relatórios"
                ],
                "summary": "Extrato detalhado da conta",
                "operationId": "72a4fcdaca649326a3d0928648428aa1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "data_inicio",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "data_fim",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Extrato da conta"
                    }
                }
            }
        },
        "/api/relatorios/clientes/{cliente}/extrato": {
            "get": {
                "tags": [
                    "Relatórios"
                ],
                "summary": "Extrato agregado de todas as contas do cliente",
                "operationId": "a75d2e6bad3b713609a0eb43bf8336ed",
                "parameters": [
                    {
                        "name": "cliente",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "data_inicio",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "data_fim",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Extrato agregado"
                    }
                }
            }
        },
        "/api/relatorios/auditoria": {
            "get": {
                "tags": [
                    "Relatórios"
                ],
                "summary": "Relatório de auditoria (logs de ações)",
                "operationId": "21d9ef1493cbfc56dc056a932e08e351",
                "parameters": [
                    {
                        "name": "data_inicio",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "data_fim",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "acao",
                        "in": "query",
                        "description": "Filtrar por ação",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Logs de auditoria"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Token de autenticação Bearer obtido via endpoint /api/login",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}